Allow not clearing composer cache

+++ php-nginx/usr/local/share/php/common_functions.sh
@@ -14,8 +14,15 @@ run_composer() {
   fi
 
   as_code_owner "composer install ${COMPOSER_INSTALL_FLAGS}"
-  rm -rf /home/build/.composer/cache/
-  as_code_owner "composer clear-cache"
+
+  do_clear_composer_cache
+}
+
+do_clear_composer_cache() {
+  if [ "${COMPOSER_CLEAR_CACHE}" == "true" ]; then
+    rm -rf /home/build/.composer/cache/
+    as_code_owner "composer clear-cache"
+  fi
 }
 
 do_composer() {
